Solve "Day 1: Report Repair"
Rating: 1 out of 5.

Honestly: I solved this in Python first, using 3 lines of code since python has an excellent standard library for these kind of tasks. itertools.combinations is exactly what this puzzle needs to get solved.

Since Elixir lacks this out of the box, and I did not feel for reverse-engineer a functionality that IMHO should "just exist", the puzzles of the first day was mainly solved by googling.

Makes me wonder if I should switch to Python instead this year.